Rink now initializes its own zones, added interfaces. Need to start accounting for collision

+++ Shootout/src/Model/Position.java
@@ -4,7 +4,7 @@ public class Position {
   private double x;
   private double y;
 
-  public Position(int x, int y) {
+  public Position(double x, double y) {
     this.x = x;
     this.y = y;
   }
